styling and fix team id value
