--- frontend/pages/SoftwarePage/components/AppStoreVpp/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AppStoreVpp";
